Refactor open code button
